Add horizontal navigation for "manage songs"
